Cap nhat logic du an, them cau hinh multi-thread cho xu ly file

+++ task-scheduler/src/main/java/com/practice/task_scheduler/services/implement/UserTaskAssignmentServiceImpl.java
@@ -109,7 +109,8 @@ public class UserTaskAssignmentServiceImpl implements UserTaskAssignmentService 
     @Override
     public void deleteAssignedUserTask(long userId, Task task) {
         UserTaskAssignment assignment = userTaskAssignmentRepository.findByTaskIdAndUserId(task.getId(), userId)
-                .orElseThrow(() -> new TaskException(ErrorCode.USERTASK_ASSIGNMENT_NOT_FOUND));
+                .orElse(null);
+        if (assignment == null) return;
         userTaskAssignmentRepository.delete(assignment);
 
         updateTaskCompletionStatusById(task.getId());
